Optimizations to Linux filepath finding for Zio
