+++ engine/Sandbox.Filesystem/BaseFileSystem.cs
@@ -596,74 +596,31 @@ public class BaseFileSystem
 	/// </summary>
 	internal string FixPath( string path )
 	{
-		if ( OperatingSystem.IsLinux() ) path = ResolveExt4FilePath( path );
-
 		// Do not allow 0-32 ASCII stuff
 		if ( path.Any( c => char.IsControl( c ) ) ) throw new ArgumentException( "Path cannot contain control characters!", "path" );
 
 		if ( path.Length < 1 )
 			return "/";
 
-		if ( path[0] == '/' ) return path;
-		return string.Concat( "/", path );
+		if ( path[0] != '/' ) path = string.Concat('/', path);
+
+		if ( OperatingSystem.IsLinux() )
+			path = ManageExt4.ResolveLinuxFilePath( system, path );
+
+		return path;
 	}
 
 	/// <summary>
-	/// Sometimes Linux filesystems cannot be resolved because the engine
-	/// attempts to compare a lowercase path to an uppercase one. This rebuilds
-	/// the folder and file paths in segments to allow the engine to build the uppercase
-	/// variant. This only applies to Zio mounts and will not fix managed calls to the native
-	/// engine.
+	/// Passes through case handled paths for Zio system root paths
 	/// </summary>
-	internal string ResolveExt4FilePath( string path )
+	/// <param name="path"></param>
+	/// <returns></returns>
+	internal string NormalizePath( string path )
 	{
-		if ( system is null ) return path;
+		path = path.Replace( "\\", "/" );
+		if ( OperatingSystem.IsWindows() ) path = path.ToLowerInvariant(); 
 
-		var segments = path.Trim( '/' ).Split( '/', StringSplitOptions.RemoveEmptyEntries );
-		if ( segments.Length == 0 ) return path;
-
-		var resolvedSegments = new List<string>( segments.Length );
-
-		foreach ( var segment in segments )
-		{
-			var parentPath = new Zio.UPath( "/" + string.Join( "/", resolvedSegments ) );
-			var candidatePath = new Zio.UPath( parentPath.FullName.TrimEnd( '/' ) + "/" + segment );
-
-			if ( system.DirectoryExists( candidatePath ) )
-			{
-				resolvedSegments.Add( segment );
-				continue;
-
-			} // If the folder exists already, do nothing.
-
-			try
-			{
-				// BaseFileSystem seems to just handle directory mounts, nothing about file names. If that changes
-				// we can adjust this to Zio.SearchTarget.Both
-				var match = system.EnumeratePaths( parentPath, "*", SearchOption.TopDirectoryOnly, Zio.SearchTarget.Directory )
-					.FirstOrDefault( p => string.Equals( p.FullName.Substring( p.FullName.LastIndexOf( '/' ) + 1 ), segment, StringComparison.OrdinalIgnoreCase ) );
-
-				var fullName = match.FullName;
-				var resolved = string.IsNullOrEmpty( fullName ) ? segment : fullName.Substring( fullName.LastIndexOf( '/' ) + 1 );
-
-				if ( resolved != segment )
-					Log.Info( $"[ResolveExt4FilePath] Corrected folder '{segment}' -> '{resolved}' (parent: {parentPath})" );
-
-				resolvedSegments.Add( resolved );
-			}
-			catch ( System.IO.DirectoryNotFoundException )
-			{
-				Log.Info( $"[ResolveExt4FilePath] DirectoryNotFound at parent '{parentPath}', keeping segment '{segment}'" );
-				resolvedSegments.Add( segment );
-			}
-		}
-
-		var result = string.Join( "/", resolvedSegments );
-
-		if ( result != path.Trim( '/' ) )
-			Log.Info( $"[ResolveExt4FilePath] '{path}' -> '{result}'" );
-
-		return result;
+		return path;
 	}
 
 	/// <summary>
@@ -673,4 +630,5 @@ public class BaseFileSystem
 	{
 		return filepath.ToLower().Replace( "\\", "/" ); // we can probably do more here
 	}
+
 }
