+++ engine/Sandbox.Filesystem/LocalFileSystem.cs
@@ -10,9 +10,8 @@ internal class LocalFileSystem : BaseFileSystem
 	internal LocalFileSystem( string rootFolder, bool makereadonly = false )
 	{
 		Physical = new Zio.FileSystems.PhysicalFileSystem();
-		rootFolder = OperatingSystem.IsLinux() ? rootFolder : rootFolder.ToLowerInvariant();
 
-		var rootPath = Physical.ConvertPathFromInternal( rootFolder );
+		var rootPath = Physical.ConvertPathFromInternal( NormalizePath( rootFolder ) );
 		system = new Zio.FileSystems.SubFileSystem( Physical, rootPath );
 
 		if ( makereadonly )
